0416
CTexture 클래스를 통해 코드 조언 및 GetRotatedBitmap 함수 구현 중

CTexture의 Load 함수는 순전히 비트맵 파일을 읽어 메모리에 할당시키기 위한 함수이기에 당장의 테스트에서도

dc 생성과 비트맵 연결을 위한 SelectObject 사이에 Rotate 함수가 들어가선 안되기에 Load가 끝났다고 가정하고 함수 마지막에서 테스트 해보고 있는데
비트맵이 정상적으로 출력되지 않고있음

+++ 20220302/20220302/Scene_Tool.cpp
@@ -41,7 +41,7 @@ void CScene_Tool::render(HDC _dc)
 	CSelectGDI pen(_dc, PEN_TYPE::GREEN);
 	CSelectGDI brush(_dc, BRUSH_TYPE::WHITE);
 
-	Rectangle(_dc, 0, 0, m_vResolution.x, m_vResolution.y);
+	Rectangle(_dc, 0, 0, static_cast<int>(m_vResolution.x), static_cast<int>(m_vResolution.y));
 
 	CScene::render(_dc);
 }
